refactor: restructure App component and modularize chart functionality
- Moved chart-related logic and types into separate files for better organization and maintainability.
- Created a new `ChartComponent` to encapsulate chart rendering and data handling, improving code clarity.
- Updated the `App` component to utilize the new structure, enhancing readability and separation of concerns.
- Adjusted the auto-refresh interval for session data from 30 seconds to 30 minutes for improved performance.
