+++ web-ui/src/App.tsx
@@ -1,163 +1,13 @@
-import {
-  // AreaSeries,
-  createChart,
-  ColorType,
-  CandlestickSeries,
-} from "lightweight-charts";
-import type { Time, ISeriesApi, IChartApi } from "lightweight-charts";
+import type { Time, ISeriesApi } from "lightweight-charts";
 import { useEffect, useRef, useState, useCallback } from "react";
-
-type Candle = {
-  time: Time;
-  open: number;
-  high: number;
-  low: number;
-  close: number;
-};
-
-type ApiCandle = {
-  close: number;
-  high: number;
-  low: number;
-  open: number;
-  timestamp: number;
-  ts: string;
-  volume: number;
-};
-
-type Balance = {
-  [symbol: string]: number;
-};
-
-type LatestSessionResponse = {
-  session_id: string;
-  candles: ApiCandle[];
-  balance: Balance;
-};
-
-// Convert API candles to chart format
-function convertApiCandlesToChart(apiCandles: ApiCandle[]): Candle[] {
-  return apiCandles.map((candle) => ({
-    time: Math.floor(candle.timestamp / 1000) as Time,
-    open: candle.open,
-    high: candle.high,
-    low: candle.low,
-    close: candle.close,
-  }));
-}
-
-// Fetch latest session data from API
-async function fetchLatestSession(): Promise<LatestSessionResponse> {
-  const response = await fetch("http://localhost:4200/session");
-  if (!response.ok) {
-    throw new Error(`HTTP error! status: ${response.status}`);
-  }
-  return response.json();
-}
-
-export const ChartComponent = (props: {
-  colors: {
-    backgroundColor: string;
-    lineColor: string;
-    textColor: string;
-    areaTopColor: string;
-    areaBottomColor: string;
-  };
-  candleData?: Candle[];
-  onSeriesReady?: (series: ISeriesApi<"Candlestick">) => void;
-}) => {
-  const { candleData, onSeriesReady } = props;
-
-  const chartContainerRef = useRef<HTMLDivElement>(null);
-
-  const chartRef = useRef<IChartApi | null>(null);
-  const seriesRef = useRef<ISeriesApi<"Candlestick"> | null>(null);
-
-  useEffect(() => {
-    if (!chartContainerRef.current) return;
-
-    const handleResize = () => {
-      if (chartRef.current) {
-        chartRef.current.applyOptions({
-          width: chartContainerRef.current?.clientWidth,
-          height: chartContainerRef.current?.clientHeight || 400,
-        });
-      }
-    };
-
-    // Only create chart if it doesn't exist
-    if (!chartRef.current) {
-      const chart = createChart(chartContainerRef.current, {
-        layout: {
-          background: { type: ColorType.Solid, color: "#000000" },
-          textColor: "#ffffff",
-        },
-        grid: {
-          vertLines: {
-            color: "#1a1a1a",
-          },
-          horzLines: {
-            color: "#1a1a1a",
-          },
-        },
-        timeScale: {
-          borderColor: "#333333",
-          timeVisible: true,
-        },
-        rightPriceScale: {
-          borderColor: "#333333",
-        },
-        width: chartContainerRef.current?.clientWidth,
-        height: chartContainerRef.current?.clientHeight || 400,
-      });
-
-      const series = chart.addSeries(CandlestickSeries, {
-        upColor: "#00a6a5",
-        downColor: "#ff1e38",
-        borderVisible: false,
-        wickUpColor: "#00a6a5",
-        wickDownColor: "#ff1e38",
-      });
-
-      chartRef.current = chart;
-      seriesRef.current = series;
-
-      // Notify parent component that series is ready for real-time updates
-      if (onSeriesReady) {
-        onSeriesReady(series);
-      }
-
-      chart.timeScale().fitContent();
-      chart.timeScale().scrollToPosition(5, true);
-    }
-
-    window.addEventListener("resize", handleResize);
-
-    return () => {
-      window.removeEventListener("resize", handleResize);
-    };
-  }, [onSeriesReady]);
-
-  // Separate effect for updating data
-  useEffect(() => {
-    if (seriesRef.current && candleData && candleData.length > 0) {
-      seriesRef.current.setData(candleData);
-    }
-  }, [candleData]);
-
-  // Cleanup on unmount
-  useEffect(() => {
-    return () => {
-      if (chartRef.current) {
-        chartRef.current.remove();
-        chartRef.current = null;
-        seriesRef.current = null;
-      }
-    };
-  }, []);
-
-  return <div ref={chartContainerRef} className="w-full h-full" />;
-};
+import type {
+  ApiCandle,
+  Balance,
+  Candle,
+  LatestSessionResponse,
+} from "./types/core";
+import { ChartComponent } from "./components/Chart";
+import { convertApiCandlesToChart, fetchLatestSession } from "./utils/core";
 
 export default function App() {
   const [sessionData, setSessionData] = useState<LatestSessionResponse | null>(
@@ -288,8 +138,8 @@ export default function App() {
   useEffect(() => {
     loadData();
 
-    // Auto-refresh every 30 seconds
-    const interval = setInterval(loadData, 60000);
+    // Auto-refresh every 30 minutes
+    const interval = setInterval(loadData, 30 * 60_000);
 
     return () => clearInterval(interval);
   }, []);
